LayoutTests/webrtc/encoded-streams-quirks.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="../resources/testharness.js"></script>
        <script src="../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="video" autoplay playsinline controls></video>
        <script src="routines.js"></script>
        <script>
promise_test(async (test) => {
    let permission = await navigator.permissions.query({ name: "camera" });
    assert_equals(permission.state, "prompt");
    permission = await navigator.permissions.query({ name: "microphone" });
    assert_equals(permission.state, "prompt");

    if (!window.internals)
        return;

    window.internals.setTopDocumentURLForQuirks("https://www.facebook.com");

    permission = await navigator.permissions.query({ name: "camera" });
    assert_equals(permission.state, "granted");
    permission = await navigator.permissions.query({ name: "microphone" });
    assert_equals(permission.state, "granted");

    const localStream = await navigator.mediaDevices.getUserMedia({audio: true, video: true});
    const stream = await new Promise((resolve, reject) => {
        createConnections((firstConnection) => {
            const audioSender = firstConnection.addTrack(localStream.getAudioTracks()[0], localStream);
            let streams = audioSender.createEncodedStreams();
            streams.readable.pipeTo(streams.writable);

            const videoSender = firstConnection.addTrack(localStream.getVideoTracks()[0], localStream);
            streams = videoSender.createEncodedStreams();
            streams.readable.pipeTo(streams.writable);
        }, (secondConnection) => {
            secondConnection.ontrack = (trackEvent) => {
                streams = trackEvent.receiver.createEncodedStreams();
                streams.readable.pipeTo(streams.writable);
                resolve(trackEvent.streams[0]);
            };
        });
        test.step_timeout(() => reject("Connections creation timed out"), 5000);
    });

    video.srcObject = stream;
    await video.play();
    await new Promise(resolve => video.requestVideoFrameCallback(resolve));
});
        </script>
    </body>
</html>
